Migrate the code from sound into video.
